add support for generic GET webhook uri
Add ability to hit a generic webhook with a GET request on a successful run.  The idea here is to setup an endpoint on healthchecks.io, which will be hit on every successful run.  If the script fails, the webhook isn't hit, which eventually triggers an alarm on healthchecks.io.

Since this is just making a simple GET request, one could also use it to hit any generic webhook on success.